+++ jest.config.js
@@ -1,3 +1,4 @@
 module.exports = {
   ...require('@jupiterone/integration-sdk-dev-tools/config/jest'),
+  setupFiles: ['dotenv/config'],
 };