+++ src/provider/types.ts
@@ -50,6 +50,7 @@ export interface GitLabGroup {
 }
 
 export interface GitLabProject {
+  _links: unknown;
   id: number;
   name: string;
   namespace: {