+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ### Fixed
 
+- [#19](https://github.com/JupiterOne/graph-gitlab/issues/19) - Deduplicate
+  `gitlab_group_has_user` relationships
 - Unhandled authentication/authorization error was not reported to user
 
 ## 2.2.0 - 2020-10-29